mmc: dw_mmc: Set maximum value at host Internal devider.
dw_mmc has Internal devider which can support up to 0xFF.
If requested clock is too small or clock source is too big
for SDR104(200Mhz), It is hard to make the optimal value
for requested clock. As a result, It is needed to change
source clock to low status sometimes.

Change-Id: Ibdf1697f9b85fae92da92a691c75f20481627e7d
Signed-off-by: jh0801.jung <jh0801.jung@samsung.com>
Signed-off-by: Hosung Kim <hosung0.kim@samsung.com>