ARM: kernel: provide cluster to logical cpu mask mapping API
Some device drivers like PMU require to retrieve the logical cpu mask
that corresponds to a given cluster id. This patch provides a hook in
the topology code that, given an existing cluster id as input,
initializes the corresponding cpumask passed as a pointer, reusing all
existing topology information required by sched domains in the kernel.

Change-Id: I256a8625f935e5d297e08e7b5058ee73e40ccaf0
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>