ARM: vfp: fix vfp context restore problem after suspend/resume
There is a problem that vfp context isn't restored
when suspend/resume test in the middle of memcpy with NEON.
So, occasionally, destination data is corrupted.
It is due to skip restoring vfp context in cpu hotplug out
according to some conditions.
This patch changes that restoring of vfp context is always occurred
in case of cpu hotplug out.

Change-Id: Ieaa183aee16c688edf390b27903bc30abcb0333b
Signed-off-by: Jun Kim <junkim@samsung.com>