usb: dwc3: core: introduce and use macros for Event Size register
That register has more than just the event buffer
size; we can also mask and unmask that particular
interrupter on bit 31 of that register.

In this patch we introduce the necessary macros
and make sure to use the new macros while also
making sure we mask interrupts during driver
removal.

Change-Id: Ida56ce4882a500b9f8de2549b2851f431cfdf762
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>