usb: dwc3: gadget: get rid of IRQF_ONESHOT
We can make use of bit 31 of the GEVNTSIZ(n)
registers to mask/unmask interrupts from that
particular interrupter.

With that feature, we can easily drop IRQF_ONESHOT
from our driver which makes it possible to
properly change IRQ priorities when using RT
patchset *and* it allows us to make use of the
scheduler to choose the proper time to handle
this IRQ thread.

Change-Id: I67999579ed62ff98fff2399182a78713750ae6fa
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>