usb: dwc3: make glue layers selectable
Glue layers are starting to have separate
requirements. For example, OMAP's glue layer
is starting to use extcon framework which
no one else needs.

In order to make it clear the proper dependencies,
we are now allowing glue layers to be selectable
so that each glue layer can list their own dependencies
without messing with the core IP driver.

Change-Id: Ie59908d68d1a23132e6f4a79bcaf57364cbf0d41
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>