usb: dwc3: gadget: drop dwc3 manual phy control
Recent versions of the core, can suspend and resume
the PHYs automatically, so we don't need to fiddle
with dwc3's Global PHY registers at all.

On versions prior to 1.94a this patch will mean
that we will never ask dwc3 to suspend the PHY.

This is an acceptable behavior or such old versions
of the core, specially considering that the only
chip known to have a version prior to 1.94a was
OMAP5 ES1.0 and that's not supported in mainline
kernel, because it was just a test spin of OMAP5.

Change-Id: I19d7579c549d0640fafd63113c3154684ec68e0d
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>