media: fimc-is: refactoring DVFS
There are some changes of DVFS logic.
The explicit scenarios are added for DVFS.
Each scenario can match one check func,
so driver determines one scenario and
asks to device what qos value is with the scenario.
It's the point of refactoring DVFS to seperate driver from device
and add more flexibility.

Change-Id: Ia5899497256f44e9229ecf39acada7e9cfad2f2a
Signed-off-by: Teahyung Kim <tkon.kim@samsung.com>