HMP: Use unweighted load for hmp migration decisions
Normal task and runqueue loading is scaled according to priority
to end up with a weighted load, known as the contribution.

We want the CPU time to be allotted according to priority, but
we also want to make big/little decisions based upon raw load.

It is common, for example, for Android apps following the dev
guide to end up with all their long-running or async action
threads as low priority unless they override the AsyncThread
constructor. All these threads are such low priority that they
become invisible to the hmp_offload routine.

Using unweighted load here allows us to maximise CPU usage in busy
situations.

[skon.hwang@samsung.com] Code sync

Change-Id: I6aab6360f86f397c565a7dc5dc2828dfd7b42103
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Hwang Seungkon <skon.hwang@samsung.com>
Signed-off-by: Seokju Yoon <sukju.yoon@samsung.com>