pwm: samsung: Fix the value of tcnt and tcmp
The TCMPB and TCNTB registers are calculated through the period
and the duty cycle. The basic calculating formula is like below
on PWM invert option:

    TCNTB = period / tin
    TCMPB = period / tin - duty / tin

However, there is a hidden tick at the end of each period on
S3C64XX and EXYNOS series. So, the formula is also changed like
below:

    TCNTB = period / tin - 1
    TCMPB = period / tin - duty / tin - 1

Additionally, PWM does not support duty cycle 0% and 100%. They
are supported with the nearest values as far as possible.

Change-Id: I57673745472752551a60efcf7f766a854655b2fd
Signed-off-by: eunki.kim <eunki_kim@samsung.com>