usb: dwc3: Introduce OTG driver for dwc3
This is first release of otg driver for the dwc3 Synopsys USB3 core.
The otg driver implements the otg final state machine and control the
activation of the device controller or host controller.

In this first implementation, only simple DRD mode is implemented,

Currently simplified OTG Final State Machine is used. In future, if
required, complete OTG FSM (phy-fsm-usb.c) can be deployed.

Change-Id: Idad1d138ed572eb649ddab5f139d554c399369d0
Signed-off-by: Anton Tikhomirov <av.tikhomirov@samsung.com>