cpufreq: interactive: stop speedchange_task when CPUFREQ_GOV_STOP
This patch stop speedchange_task when CPUFREQ_GOV_STOP.
The speedchange_task can be accessed to policy when policy cpu
is hot plugged out. In this case, the policy has already been
removed. So, speedchange_task should be stop when CPUFREQ_GOV_STOP,
and should be start when CPUFREQ_GOV_START.

Change-Id: I474c0fb77487c5be05f7d89e4f5a98be98cd2b21
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>