[COMMON] sched: make __update_entity_runnable_avg() more accurate
delta_w in __update_entity_runnable_avg() is
already accumulated value against our next period.
Thus, delta_w of next period should be same as remainder of delta
in previous period but in current implementation, this is not the case.

To solve this problem, this patch saves remainder of delta and
uses it as delta_w in next period. Additionally checks whether
runnable is set before updating delta_w because remainder of delta
in previous period is meaningful only if the thread is runnable state.

Change-Id: I6fd42e1508c0571d5eb1703fbc0c6ca78ddeb5e7
Signed-off-by: Hwang Seungkon <skon.hwang@samsung.com>
Reviewed-by : Donghee Han <dh.han@samsung.com>
[updated changelog and update delat_w only if runnable is set]
Reviewed-by : Changhwan Youn <chaos.youn@samsung.com>