THERMAL: fix for kernel panic due to null pointer dereference
There was a kernel panic in the 10/30 reboot aging test regarding
DTM driver. This was due to the early access on cpufreq policy.governor
before cpufreq was brought up. Previously in the DTM driver, cpufreq
policy was updated after null check on policy only. However, the symptom
seems to happen due to the null value on policy.governor.
Therefore, null check on policy.governor is added before policy update.

Change-Id: I396a86fe3a367884392fbfb62f8e23daaba940cb
Signed-off-by: Jang JeongHoon <jnghn.jang@samsung.com>