[3250] ARM: add atomic_push() and atomic_pop()
atomic_push() and atomic_pop() is useful for managing small amount
of small values with stack data structure.
The size of the stack is defined by the bit-width of a vaule that
will be stored in the stack.
If the range of the value is 1 ~ 5, and the bit-width of each vaule
is 4, the stack size is 8 in 32-bit system because the size of atomic
variable is 32-bit.

Change-Id: Iecfb3e4e5d97b2742b861c21325856f4be1c3167
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>