media: scaler: fix IO virtual memory leak while downscaling under 1/4
Due to the specification of Scalers that do not support donw-scaling
under 1/4, scaling down to 1/16 is supported by the Scaler driver that
executes Scalers twice inside the driver. The 'output' buffer from
the user is the output buffer of the first execution and the 'capture'
buffer from the user is the capture boffer of the second execution.
The driver needs intermediate buffers that works as the capture buffer
of the first execution and the output buffer of the second execution.
Therfore the driver allocates intermediate buffer from ION and maps
the allocated buffers to the IO virtual address space with IOVMM.

The memory leak is happened here that the driver allocates the
intermediate buffers every frame but it is freed when the user calls
reqbufs(). This patch fixes the problem with moving initialization and
cleaning up of the intermediate buffers to the right position.

Change-Id: Id0d018f00661dcbfcbaf7a5dbf1d64f75b887178
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>