[COMMON] gpu: ion: introduce ION_FLAG_NOZEROED
ION_FLAG_NOZEROED in flags argument to ion_alloc() informs ION that
the allocated buffer does not need to be initialized with zero. It
also implies that the allocated buffer will be never accessed from
userspace. Any request of mmap() of a buffer with ION_FLAG_NOZEROED
will always fail because userspace must not access non-initialzed
buffer.

Change-Id: I198f2854983bd1e290a9faab609b6eee0c8b331e
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>