[COMMON] iommu/exynos: refector exynos-iommu driver series
Two versions of exynos-iommu driver were there: one is exynos5-iommu
for System MMU version 1, 2 and 3, and the other is exynos7-iommu for
System MMU version 5 and higher. Those System MMUs are very similar
but detailed configurations. Thus most of the driver code and logic
are the same.
This patch gathers most of the code that can be shared between all
System MMU versions and prepares later System MMU versions and
enahancements that can be applied to all versions of System MMUs.

Change-Id: I89e01c771666836a2937c6f77dfb84cf464217dc
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>