[5430] PM: APM: Add clock save/resore and reconfiguration of MPHY_PLL at suspend/resume
To reduce power consumption, MPHY_PLL should be off because this PLL is not used at normal.
But, if MPHY_PLL is off, it fails to enter suspend mode currently.
To solve this, add save/restore ENABLE_IP_CPIF0 related to MPHY_PLL and
it change from 'gate' to 'pass' before suspend mode.
After resume, it returns to previous setting value.

Change-Id: If474c2a7f77b8ccbccc80ceed6ccb10e26b67a78
Signed-off-by: SangWook Ju <sw.ju@samsung.com>