[COMMON] THERMAL: remove e-fuse limit on ares2
There is an issue where temperature of specific TMU
is lower than others. It is due to the default e-fuse
value which is used when the EFUSE value is out of
the valid range. According to guide from AP dev team,
this limit for EFUSE value is not needed for ares2, so
featuring has been added.

Change-Id: I191faa0821ee13dae21639ee095628f5eafa7774
Signed-off-by: Jang JeongHoon <jnghn.jang@samsung.com>