[COMMON] iommu/exynos: promote driver registration to arch_initcall
Some device drivers like seiren are registered by subsys_initcall()
that is called earlier than subsys_initcall_sync() and the devices
that are probed before subsys_initcall_sync() are not serviced by
exynos-iommu.

This patch promotes the timing of the exynos-iommu driver registration
to arch_initcall() and arch_initcall_sync().

Change-Id: I7ba57c32175d150778970a37d59a7098f19e4ba0
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>