[COMMON] cpufreq: cpufreq-mp: bug fix for max lock function
if the clamp is above the current frequency you shouldn't return early
because you want increase the current frequency if it was clamped
by a previous clamp from IPA.
For example: Time 0 : Clamp = 50, Gov request = 100 => Current Freq will
be 50 because of clamp Time 1: IPA relaxes clamp value to 80.
Now Current Freq should be allowed to go up to 80. But those lines
imply an early exit with no change in frequency.

Change-Id: I52e0c0626c421e2db87221141b954df1d5928900
Signed-off-by: Jongpill Lee <boyko.lee@samsung.com>