[COMMON] cma: keep requested contiguous size
CMA rounds up the contiguous memory size to be page-block size aligned
or MAX_ORDER aligned. If the required contiguous memory size is not
aligned by them, isolation is performed that is not used.

This commit keeps the requested contiguous memory size so that the
isolation is performed on the exact size required.

Change-Id: I17252127e1be0989cf7517d6c0592d233ea5e57d
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>