[5430] misc: mipi-lli: Change check condition for LPA/DSTOP.
LLI was suspended independently when AP is wakeup or not.
In this case, there is no chance to go to LPA/DSTOP when
LLI clock is not gated.
This patch makes AP goes to LPA/DSTOP if LLI is unmounted.
And Modem I/F will not make LLI goes suspend when AP is wakeup.

Change-Id: Icddc6ad26a9fc753350383a6834f72f089b1839d
Signed-off-by: Byeongjo Park <bjo.park@samsung.com>