[DO NOT RELEASE][COMMON] ARM: EXYNOS: ESS: update exynos-snapshot v1.5
This patch is major changed version of exynos-snapshot as v1.5
The whole structure of exynos-snapshot is changed for scalabillty.
The following are shown the updated features. I want to leave more
evidence to grab the system lockup or core lockup called hang.
And it is available to use sysfs for control at runtime, sysfs path is
/sys/devices/system/exynos-ss

- Support to trace IRQ disable/enable period and IRQ exit event
- Support to trace all accessed registers
- Support to trace kernel event with callstack at runtime
- Support to restore panic at the exact more
- Support various options for developers and some bug fixes
- Support sysfs to control the various options
- Support all trace with on/off function at runtime

Change-Id: Id2e05131dc8c685fd487e596058ff42b5e06de13
Helped-by: Jinsung Yang <jsgood.yang@samsung.com>
Signed-off-by: Hosung Kim <hosung0.kim@samsung.com>