[543x] video: exynos: decon: Remove DECON_DUMMY_WIN_DISPLAY feature
DMA operation must be needed while disabling decon's IP clock.
So, at least, one window was working at that time.

However, this W/A is changed to gate MUX on behalf of gating IP clock.
As a result, previous W/A that is DECON_DUMMY_WIN_DISPLAY feature is removed.

Change-Id: Id01f1718b96b31c223ce1805d2a0da207e3c0f23
Signed-off-by: SeWoon Park <seuni.park@samsung.com>