[COMMON] media: scaler: optimize sc_fmt
sc_fmt has some fields that just stores several bits. It is an array
of a number of color formats and optimizing the size of fields saves
a hundred bytes.
Additionally, the array of sc_fmt is restricted by const qualifier
to notice the code readers that the structure is never changed.

Change-Id: Id8fe3a9191588bad9e8d284c61ff6d28a3642485
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>