arm64: topology: Tell the scheduler about the relative power of cores
In heterogeneous systems like big.LITTLE systems the scheduler will be
able to make better use of the available cores if we provide power numbers
to it indicating their relative performance. Do this by parsing the CPU
nodes in the DT.

This code currently has no effect as no information on the relative
performance of the cores is provided.

Change-Id: Ide6e450c9199e71444c49b67a71cdf7cd3ac589e
Signed-off-by: Mark Brown <broonie@linaro.org>
Signed-off-by: Seokju Yoon <sukju.yoon@samsung.com>