[COMMON] media: scaler: fix violation of restriction in resolution
Scaler driver supports x1/16 scaling even though H/W just supports
x1/4 scaling with double excuetions. The basic algorithm of the
double execution is scaling down the original image to x1/4 scale
then process the rest of the scaling ratio.
However, the first output of the x1/4 scaling may violate the
restriction of H/W specification of the source image resolution.
51 -> 4 is the case, for example. 51 / 4 is 12.x but the smallest
width or height of the source image is 16.
This patch restricts the width or the height of the first output
image not to violate H/W restrictions. 51 -> 4 scaling will be
51 -> 16 -> 4.

Change-Id: I8a3458804a2dbd6e04ef4b708bd4fc73af7d0f9a
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>