[3250] fimc-is-w1: Add ISP ARM power off code
3.4 kernel relies on power domain to take care of powering on/off
the ISP ARM. But same is not valid in 3.10 kernel as the ISP ARM
configuration and status registers have different bits compared
to other generic power domains.
This change can be taken care in the power domain itself with
modified handling for specific domains. Till then this patch
be used in the driver to take care of powering off the ISP ARM.
Strangely powering ON the ISP ARM is directly done in the driver
without using the power domain path.

Change-Id: Ibcf953b66d116a0c13eb36c6057ad90ef5c365d0
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Reviewed-by: Pankaj Kumar Dubey <pankaj.dubey@samsung.com>