[3250] fimc-is-w1: fimc-is power on error recovery
During some error scenarios in ageing test, the ISP ARM power off
may not happen. This causes the camera to stop working as the next
time power ON will not give the startup interrupt from IS.
So this patch turns off ISP ARM before turning it ON so that the
previous state doesnt matter. This is seen to significantly
improve the long run stability of camera.

Change-Id: Id8188a7763b1d9f1cfe98cfae69124d571df3a14
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Reviewed-on: http://107.108.3.115:8081/4823
Reviewed-by: Pankaj Kumar Dubey <pankaj.dubey@samsung.com>
Reviewed-by: ritesh solanki <r.solanki@samsung.com>