[3250] mipi-dsi: Use common phy control code
The MIPI phy control is a common register controlling both
MIPI CSIS and MIPI DSIM. So when decon mipi is turned off,
it disables the phy unconditionally causing camera to malfunction.
This patch removes phy enable / disable function from mipi dsi
driver and calls the common function which will be now used
by both mipi csi and dsi thereby taking care of disabling phy
only when both are disabled.
This fixes the s2r issue when camera preview is on.

Change-Id: Ia283a6786b1357a2a21f5f6019cb259b78701d24
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Reviewed-on: http://107.108.3.115:8081/4918
Reviewed-by: Naushad K <naushad@samsung.com>
Reviewed-by: Pankaj Kumar Dubey <pankaj.dubey@samsung.com>