[3250] ARM: dts: Update correct dvfs values for fimc-is
Update the values for DVFS for fimc-is. The levels are selected
as per the DVFS table and same levels are used for all scenarios.
This can be changed as and when more inputs for DVFS levels are
available. The patch also removes the unused properties like
front cam, dual camera etc which are not supported in 3250.

Change-Id: Id161a3e8fe244f14ab58c7246f850daf92c5dc63
Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Signed-off-by: Tony K Nadackal <tony.kn@samsung.com>
Reviewed-on: http://107.108.3.115:8081/4924
Reviewed-by: Tony
Tested-by: Tony
Reviewed-by: Pankaj Kumar Dubey <pankaj.dubey@samsung.com>