usb: gadget: Fix for USBCV 150 iterations of enumeration
USBCV 2.0 chapter 9 tests failed because of 150 iteration
of enumeration test failed after single iteration. USB packet
analyzer showed that SET_ADDRESS request is failing giving TURN
around timed out.
On debugging we found that SET_ADDRESS request is properly handled
by driver, but still host PC is not able to ACK for setup packet.
Also observed that it's a timing issue, and putting a delay of 50 msec
solves this issue.
So putting a delay of 50 msec to resolve this issue.

Change-Id: I2ea4bc87ff5b09466a168a8eda92ecd4533b4619
Signed-off-by: Pankaj Dubey <pankaj.dubey@samsung.com>
Reviewed-on: http://107.108.3.115:8081/5048
Reviewed-by: Praveen Paneri <p.paneri@samsung.com>
Reviewed-by: ritesh solanki <r.solanki@samsung.com>