[7420] media: scaler: compromise pre-scaling restriction
Pre-scaler merges two or four pixels into one pixel.
The result resolution should also satisfy the restriction of the
image format.

H/W requires that the pre-scaler input resolution should be twice
or 4 times larger than the result resolution according to the
pre-scaling ratio.

The scaler may also support the source image resolutions which do not
satisfy the restriction of the H/W. This patch forces < x1/4
down-scaling with double execution of H/W if the input image resolution
violates the restriction of pre-scaler.

Change-Id: Ifc5965f19c9e9ffd8339144689a8fb56caaed8a8
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>