dmaengine: split out pause/resume operations from device_control
[ Upstream commit 23a3ea2f5bead4d3b16e119e9127a66234f41d53 ]

Split out the pause and resume operations to callbacks of their own. In order
to preserve some backwark compatibility, the dmaengine_pause/dmaengine_resume
are still falling back on dmaengine_device_control.

Eventually, that will allow to get the device capabilities in a generic way,
removing the need to implement device_slave_caps.

Change-Id: I49c1edc77aacc76ddb60239551f5b1dba605eed6
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>
Acked-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>