[COMMON] thermal: fix for kernel panic due to null pointer dereference
There was a kernel panic in the ACT test when DCH is enabled.
This was due to the early access on cpufreq policy.governor
before cpufreq was brought up. Previously in the DTM driver, cpufreq
policy was updated after null check on policy only. However, the symptom
seems to happen due to the null value on policy.governor.
Therefore, null check on policy.governor is added before policy update.

Cc: Jang JeongHoon <jnghn.jang@samsung.com>
Cc: Taekki Kim <taekki.kim@samsung.com>
Signed-off-by: Jonghwan Choi <jhbird.choi@samsung.com>

Conflicts:

	drivers/thermal/samsung/cpu_cooling.c

Change-Id: I1ce92bfd5d1008eac3210cbf77032327d9d1c142