rtc-sec: workaround for the RTC TIME overwrite
This is a workaround for the problem that RTC TIME is overwirted by write
buffer when setting RTC ALARM. It is quite rare but it does happen. The root
cuase is that clear signal of RUDR & WUDR is 1 clock delay while it should be
2 clock delay.
This workaround is not dependent on system time.

Bug:62433085

Change-Id: I19f034be6cf6e2c1181cf45be70a31d5d0166b39
Signed-off-by: Ritesh Kumar Solanki <r.solanki@samsung.com>
Reviewed-on: http://107.108.3.115:8081/6464
Reviewed-by: Pankaj Kumar Dubey <pankaj.dubey@samsung.com>
Tested-by: Pankaj Kumar Dubey <pankaj.dubey@samsung.com>