[COMMON] media: mfc: fix the release sequence
If there is a gap between s5p_mfc_clean_ctx_int_flags
and set_bit of ctx_work_bits, we can lose the result
that it is performed by the try_run of the other instance.
Also if we judged that HW state is abnormal,
we have to return the clock ref_count normally.

Change-Id: I1674923d954ff44440e30b6fac8bbc463fdd6b95
Signed-off-by: Ayoung Sim <a.sim@samsung.com>