[COMMON] media: m2m1shot: accept default value of bytes_used.
bytes_used is the value decided by the drivers according to the image
format provided by the users. m2m1shot reports an error to the users
if the length of the buffer is smaller than bytes_used calculated by
the drivers. However, there is a case that the drivers are not able to
decide bytes_used because the image is not a raster image like JPEG
compressed stream.

Threfore m2m1shot now allows the drivers to leave bytes_used as its
default value, zero. m2m1shot then sets it to the length of the buffer.

Change-Id: I0c3e2f218e525bdf3913fe4e506fc00c07b29be8
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>