ext4: insert 3.18 version of fs/ext4, fs/jbd2 and associated header files
This will cause ext3 and gfs2 to not compile correctly, but allows us
to get a modern version of ext4 into 3.10.  This makes it easier to
backport newer features such as ext4 encryption into a downrev kernel.
It also fixes a number of xfstest failures that were fixed since 3.10.

The subsequent commits will fix up the 3.18 ext4 codebase so it will
compile against 3.10.

Signed-off-by: Theodore Ts'o <tytso@google.com>

Conflicts:

	fs/ext4/ext4_jbd2.c
	fs/ext4/extents.c
	fs/ext4/inode.c
	fs/ext4/super.c

Change-Id: Iad750b375d49780ed1074b1ba1a1ca34555fd1d3