[COMMON] media: mfc: support BUFFER_FULL interrupt
Since host does not know the exact size
of encoded bitstream in advance,
a stream buffer (E_STREAM_BUFFER_ADDR) may not be enough
to accommodate if E_STREAM_BUFFER_SIZE is small.
When the stream buffer is full but picture encoding is not completed,
MFC raises a RET_ENC_BUFFER_FULL interrupt.
When host receives this interrupt, there is a possibility
that host would like to stop encoding and discard the current picture.

Change-Id: Ifadfb7e5b1a350fc2329f450f2da3cc405918442
Signed-off-by: Ayoung Sim <a.sim@samsung.com>