[common]: gpio: Correct pin configuration for BT
When in Ambient-sleep mode, Bluetooth should not be turn OFF,
to prevent this we need to keep BT power alway ON. BT power is controlled
by pwr_en gpio. This patch correct the setting of these gpios, so that
power to BT can be always ON.
This patch also remove the GPX which are part of Alive block and should not
be controlled here.

Change-Id: I7cdaeb03fcc0a680a12f25f3dad34c7618542c35
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>