[COMMON] iommu/exynos: remove unused plane number for iova allocation
A plane number or an window number is required to decide which part of
a IO virtual address space is involved in a memory allocation by IOVMM
to avoid possible H/W hang by prefetch buffers of System MMU 3.x.

Exynos5433 is the last SoC that embeds System MMU 3.x and Linux kernel
3.18 does not supports for Exynos5433. Therefore, it is okay to simply
remove the obsolete region ID support.

Change-Id: I07646d4b9f423c70f79eb0056a18793ab700d5ca
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>