[COMMON] media: vb2-ion: remove obsolete vb2_ion_set_protected()
vb2_ion_set_protected() switches the address space of the specified
device between the virtual address space and the physical address
space. It is required when a System MMU works as a secure contents
firewall. But vb2_ion_set_protected() is no longer valid since
Exynos7420 because separate firewalls are installed in every SoC.

Change-Id: I7cfa25fa5267edae9ad203f423a2a5ea187ffbd4
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>