ARM: exynos: prevent mct-comp timer from being set in the past
tick_broadcast_oneshot_control eventually calls through to
clockevents_program_event with force set, which may attempt
to program the mct-comp broadcast timer with the minimum
number of cycles allowed for the timer, which is currently
set to 0xf.  If the timer counts more than 0xf cycles while it
is being programmed, the interrupt will never fire.
tick_broadcast_oneshot_control ignores the return value, so
returning an error won't help.  Increase the minimum number of
cycles to 0xff to make sure the timer can't increase past the
target being programmed, and also turn of interrupts during
the programming to make sure that we don't get preempted during
that region.

Change-Id: Iaf3af7a386f24b6a4158185021b2c94b017ded3d
Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: Jiseong Oh <jiseong.oh@samsung.com>