[COMMON] scsi: ufs-exynos: move some controls from hibernation to clock control
In some SoCs, there are two limitations that are not compatible
with clock control in UFS host driver.
1. When UFS host is in hibernation state, system can enter into
system power mode at any time that UFS host don't know,
that means not using notifications
2. Disabling clocks using clk_disable/enable should be used
only when the clocks are turned on. Because the disabling/enabling
in the SoC means switching clock sources and the clocks
enter unstable states when they are turned off

Hence, operations to block clock sources internally should be not
included in hibernation controls. So they are moved into
clock controls like following:
(hibernation enter) -> clock disabling -> interal disabling -> ...
-> internal enabling -> clock enabling -> (hibernation exit)

Change-Id: I9e28ca41d1346d6c824c394d96e74f70751486ba
Signed-off-by: Kiwoong Kim <kwmad.kim@samsung.com>