[COMMON] scsi: ufs: add power change notify callback
This change adds support for the vendor specific host driver to
implement power change notify callback.

To enable configuring different power modes (number of lanes,
gear number and fast/slow modes) it is necessary to split the
configuration stage from the stage that reads the device max power mode.
In addition, it is not required to read the configuration more than
once, thus the configuration is stored after reading it once.

Change-Id: If9f0c4eccd56a33b11e61adc25d4346663bff9fe
Signed-off-by: Seungwon Jeon <tgih.jun@samsung.com>
Signed-off-by: soojung kang <sj88.kang@samsung.com>